requirements.txt
sqlparse==0.4.1
static3==0.7.0
whitenoise==5.2.0
psycopg2==2.8.5
django-environ==0.4.5
gunicorn==20.0.4